sum of digit, table a number, tax shortcode, total of numbers, prime numbers, square and cube of number, prime number, one to ten number print, factorial number

+++ tax.py
@@ -1,5 +1,5 @@
-totalIncome = int(input("Your Income: "))
-saving = int(input("Your saving: "))
+totalIncome = float(input("Your Income: "))
+saving = float(input("Your saving: "))
 taxIncome = totalIncome - saving
 taxIn = totalIncome - 150
